requirements.txt
playwright==1.49.1
requests==2.31.0
python-dotenv==1.0.0
